// old/packages/waas/src/clients/intent.gen.ts
export enum IntentResponseCode {
  authInitiated = 'authInitiated',
  sessionOpened = 'sessionOpened',
  sessionClosed = 'sessionClosed',
  sessionList = 'sessionList',
  validationRequired = 'validationRequired',
  validationStarted = 'validationStarted',
  validationFinished = 'validationFinished',
  sessionAuthProof = 'sessionAuthProof',
  signedMessage = 'signedMessage',
  signedTypedData = 'signedTypedData',
  feeOptions = 'feeOptions',
  transactionReceipt = 'transactionReceipt',
  transactionFailed = 'transactionFailed',
  getSessionResponse = 'getSessionResponse',
  accountList = 'accountList',
  accountFederated = 'accountFederated',
  accountRemoved = 'accountRemoved',
  idToken = 'idToken',
  adopter = 'adopter',
  childWalletAdopted = 'childWalletAdopted',
  confirmationRequired = 'confirmationRequired'
}

export enum FeeTokenType {
  unknown = 'unknown',
  erc20Token = 'erc20Token',
  erc1155Token = 'erc1155Token'
}

export enum IdentityType {
  None = 'None',
  Guest = 'Guest',
  OIDC = 'OIDC',
  Email = 'Email',
  PlayFab = 'PlayFab',
  Stytch = 'Stytch'
}
// ...
export interface IntentDataSendTransaction {
  network: string
  wallet: string
  identifier: string
  transactions: Array<any>
  transactionsFeeQuote?: string
}
// ...
export interface IntentResponseAuthInitiated {
  sessionId: string
  identityType: IdentityType
  expiresIn: number
  challenge?: string
}

export interface IntentResponseSessionOpened {
  sessionId: string
  wallet: string
}

export interface IntentResponseSessionClosed {}

export interface IntentResponseValidateSession {}

export interface IntentResponseValidationRequired {
  sessionId: string
}

export interface IntentResponseValidationStarted {
  salt: string
}

export interface IntentResponseValidationFinished {
  isValid: boolean
}

export interface IntentResponseListSessions {
  sessions: Array<string>
}

export interface IntentResponseGetSession {
  sessionId: string
  wallet: string
  validated: boolean
}

export interface IntentResponseSessionAuthProof {
  sessionId: string
  network: string
  wallet: string
  message: string
  signature: string
}

export interface IntentResponseSignedMessage {
  signature: string
  message: string
}

export interface IntentResponseSignedTypedData {
  signature: string
  encodedTypedData: string
}

export interface FeeOption {
  token: FeeToken
  to: string
  value: string
  gasLimit: number
}

export interface FeeToken {
  chainId: number
  name: string
  symbol: string
  type: FeeTokenType
  decimals?: number
  logoURL: string
  contractAddress?: string
  tokenID?: string
}

export interface IntentResponseFeeOptions {
  feeOptions: Array<FeeOption>
  feeQuote?: string
}

export interface IntentResponseTransactionReceipt {
  request: any
  txHash: string
  metaTxHash: string
  receipt: any
  nativeReceipt: any
  simulations: any
}

export interface IntentResponseTransactionFailed {
  error: string
  request: any
  simulations: any
}

export interface IntentResponseAccountList {
  accounts: Array<Account>
  currentAccountId: string
}

export interface IntentResponseAccountFederated {
  account: Account
}

export interface IntentResponseAccountRemoved {}

export interface IntentResponseIdToken {
  idToken: string
  expiresIn: number
}

export interface IntentResponseChildWalletAdopted {
  adopterAddress: string
}

export interface IntentResponseAdopter {
  adopterAddress: string
}

export interface IntentResponseConfirmationRequired {
  confirmationId: string
  salt: string
  challengeType: ChallengeType
  challengeDestination?: string
  expiresIn: number
}

export interface Account {
  id: string
  type: IdentityType
  issuer?: string
  email?: string
}